installation with git needs composer, but installation with composer does not.

+++ firstrun.php
@@ -2,5 +2,8 @@
   $msg = PHP_EOL.PHP_EOL;
   $msg .= 'Thanks for trying out Coldreader!'.PHP_EOL;
   echo($msg);
-  exec('composer install --no-dev --no-suggest');
+  if (!file_exists('./vendor/imonroe/ana/src/Ana.php')){
+    echo('Looks like you installed via git. First we will get the composer dependencies.'.PHP_EOL.PHP_EOL);
+    exec('composer install --no-dev --no-suggest');
+  }
 ?>
